core/pyproject.toml
[project]
name = "ai-agents-core"
version = "0.1.0"
description = "Shared utilities for AI agents"
requires-python = ">=3.11"
dependencies = [
    "google-adk>=1.26.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
